requirements.txt
fastapi
uvicorn
redis
celery
python-jose[cryptography]
passlib[bcrypt]
slowapi
loguru
prometheus-fastapi-instrumentator
pydantic